Create README.md
Hi, I am excited to join GitHub and learn as much as possible. Besides going to school, I like Woodburn, remodeling the house, And remodeling RVs. I am currently renovating a 1987 Fan. I took it down to the shell and built it from the frame up. I plan to live in it and travel during the winter.  I love animals, being out in the fresh air working in the yard or working on cars, building my grandson a village of his own in the backyard.  I stay busy when emailing, please keep that in mind, and I will get back to you as soon as possible.
Thank you,
 Auntie-TT